Moved component files to where they belong
